refactor(gui): remove redundant layout from `GameState`, restructure HUD initialization
- Removed `layout` field from `GameState` and transitioned HUD to use a global `Gui_Layout`.
- Simplified `Draw_Hud` logic to ensure modularity and avoid reinitialization of layout.
- Adjusted `Draw_Night_Overlay` call in `main.odin` for type consistency.
- Deleted unused night overlay procedure and its logic from `gui/world.odin`.

+++ main.odin
@@ -32,7 +32,7 @@ main :: proc() {
 		rl.BeginDrawing()
 		rl.ClearBackground(gui.sky_color(game_tick))
 		gui.Draw_World(&state)
-		gui.Draw_Night_Overlay(game_tick, gui.SCREEN_W, gui.SCREEN_H)
+		gui.Draw_Night_Overlay(i32(game_tick), gui.SCREEN_W, gui.SCREEN_H)
 		gui.Draw_Hud(&state)
 		rl.EndDrawing()
 	}
